fix: openapi schemas brought up to date

+++ projects/api/src/contracts/users/schema/request/coverRequestSchema.ts
@@ -2,5 +2,5 @@ import { z } from '../../../_internal/z.ts';
 
 export const coverRequestSchema = z.object({
   cover_type: z.enum(['movie', 'show', 'episode']),
-  cover_id: z.number(),
+  cover_id: z.number().int(),
 });
